+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   ownership to be marked in the same zone for delegation NS records.
 * octodns-report access --lenient flag to allow running reports with records
   sourced from providers with non-compliant record data.
+* Correctly handle FQDNs in TinyDNS config files that end with trailing .'s
 
 ## v1.0.0.rc0 - 2023-05-16 - First of the ones